#32 - dao call is implemented in service level and view in controller in Login controller
