+++ src/com/lftechnology/java/training/alina/jdbc/controller/LoginController.java
@@ -1,14 +1,22 @@
 package com.lftechnology.java.training.alina.jdbc.controller;
 
+import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.Scanner;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
 import com.lftechnology.java.training.alina.jdbc.constants.Constants;
-import com.lftechnology.java.training.alina.jdbc.dao.user.impl.UserDaoImpl;
 import com.lftechnology.java.training.alina.jdbc.dbutils.DbFacade;
 import com.lftechnology.java.training.alina.jdbc.demo.EmployeeManagement;
+import com.lftechnology.java.training.alina.jdbc.domain.Employee;
+import com.lftechnology.java.training.alina.jdbc.domain.User;
+import com.lftechnology.java.training.alina.jdbc.routes.UserRoutes;
+import com.lftechnology.java.training.alina.jdbc.service.UserService;
+import com.lftechnology.java.training.alina.jdbc.service.UtilityService;
+import com.lftechnology.java.training.alina.jdbc.views.AdminView;
+import com.lftechnology.java.training.alina.jdbc.views.EmployeeView;
+import com.lftechnology.java.training.alina.jdbc.views.LoginView;
 
 /**
  * LoginController consists of login,logout functionals
@@ -29,9 +37,10 @@ public class LoginController {
     public void userLogin() throws SQLException {
         Boolean isLogin = false;
         try (Scanner scanner = new Scanner(System.in)) {
-            UserDaoImpl userDao = new UserDaoImpl();
+            UserService userService = new UserService();
             do {
-                isLogin = userDao.checkEmployeeLogin(scanner);
+                User user = LoginView.displayLoginHeader(scanner);
+                isLogin = userService.checkEmployeeLogin(scanner, user);
             } while (!checkNotLogin(isLogin));
         } catch (Exception e) {
             LOGGER.log(Level.WARNING, "Exception Message : {0}", new Object[] { e });
@@ -58,28 +67,52 @@ public class LoginController {
     }
 
     /**
-     * Checks existence of user by username
+     * Gets Admin role functions
      * 
      * @param scanner
      *            {@link Scanner}
-     * @param userExist
-     *            {@link Boolean}
-     * @param username
-     *            {@link String}
-     * @return userExist {@link Boolean} identifies user exists/not
+     * @param result
+     *            {@link ResultSet}
      * @throws SQLException
      * @author Alina Shakya <alinashakya@lftechnology.com>
      */
-    public static boolean checkExistUsername(String username) throws SQLException {
-        UserDaoImpl userDao = new UserDaoImpl();
-        boolean checkUserExist = userDao.checkValidUserByUsername(username);
-        if (checkUserExist) {
-            LOGGER.log(Level.INFO, Constants.USERNAME_ALREADY_EXISTS);
+    public static void getAdminRole(Scanner scanner, Employee employee) throws SQLException {
+        UserRoutes userRoutes = new UserRoutes();
+        char choice = ' ';
+        while (choice != Constants.ADMIN_EXIT) {
+            AdminView.displayAdminRoleMenu(employee);
+            choice = UtilityService.getSelectedMenu(scanner, Constants.SELECT_ADMIN_OPTION);
+            userRoutes.getAdminRoleOptions(scanner, choice);
         }
-        return checkUserExist;
     }
 
-    public static void logoutPage() throws SQLException {
+    /**
+     * Gets role of normal user
+     * 
+     * @param scanner
+     *            {@link Scanner}
+     * @param result
+     *            {@link ResultSet}
+     * @throws SQLException
+     * @author Alina Shakya <alinashakya@lftechnology.com>
+     */
+    public static void getNormalUserRole(Scanner scanner, Employee employee) throws SQLException {
+        UserRoutes userRoutes = new UserRoutes();
+        char choice = ' ';
+        while (choice != Constants.EMPLOYEE_EXIT) {
+            EmployeeView.displayEmployeeRoleMenu(employee);
+            choice = UtilityService.getSelectedMenu(scanner, Constants.SELECT_NORMAL_USER_ROLE);
+            userRoutes.getEmployeeRoleOptions(scanner, choice, employee);
+        }
+    }
+
+    /**
+     * Logouts user
+     * 
+     * @throws SQLException
+     * @author Alina Shakya <alinashakya@lftechnology.com>
+     */
+    public static void logout() throws SQLException {
         LoginController login = new LoginController();
         login.userLogin();
     }
