+++ src/com/lftechnology/java/training/alina/jdbc/views/LoginView.java
@@ -1,7 +1,13 @@
 package com.lftechnology.java.training.alina.jdbc.views;
 
+import java.sql.SQLException;
+import java.util.Scanner;
 import java.util.logging.Level;
 import java.util.logging.Logger;
+
+import com.lftechnology.java.training.alina.jdbc.constants.Constants;
+import com.lftechnology.java.training.alina.jdbc.domain.User;
+import com.lftechnology.java.training.alina.jdbc.service.UserService;
 
 public class LoginView {
 
@@ -11,7 +17,10 @@ public class LoginView {
 
     }
 
-    public static void displayLoginHeader() {
+    public static User displayLoginHeader(Scanner scanner) throws SQLException {
+        User user = new User();
         LOGGER.log(Level.INFO, "\n==============================\nLOGIN PAGE\n==============================\n");
+        user = UserService.setLoginInfo(scanner, Constants.USER_LOGIN);
+        return user;
     }
 }
